refactor(SoC): 책임에 따른 함수 관심사 분리
